remove old repos table
