+++ packages/web/src/app/(app)/repos/page.tsx
@@ -1,185 +1,130 @@
-import { sew } from "@/middleware/sew";
-import { ServiceErrorException } from "@/lib/serviceError";
-import { isServiceError } from "@/lib/utils";
-import { withOptionalAuth } from "@/middleware/withAuth";
+import { getBullMQClient } from "@/lib/bullmqClient";
+import {
+    authenticatedPage,
+    type OptionalAuthOptions,
+} from "@/middleware/authenticatedPage";
+import {
+    REPO_INDEX_QUEUE,
+    type WorkloadJob,
+} from "@sourcebot/shared";
+import { OrgRole, type Prisma } from "@sourcebot/db";
+import { z } from "zod";
 import { ReposTable } from "./components/reposTable";
-import { RepoIndexingJobStatus, Prisma } from "@sourcebot/db";
-import z from "zod";
-
-const numberSchema = z.coerce.number().int().positive();
 
 const DEFAULT_PAGE_SIZE = 20;
+const pageSchema = z.coerce.number().int().positive();
+const sortBySchema = z.enum(["name", "indexedAt"]);
+const sortOrderSchema = z.enum(["asc", "desc"]);
+const statusSchema = z.enum(["failed", "warning"]);
 
-interface ReposPageProps {
+type ReposV2PageProps = {
     searchParams: Promise<{
         page?: string;
-        pageSize?: string;
         search?: string;
         status?: string;
         sortBy?: string;
         sortOrder?: string;
     }>;
-}
+};
 
-export default async function ReposPage(props: ReposPageProps) {
-    const params = await props.searchParams;
+export default authenticatedPage<
+    ReposV2PageProps,
+    OptionalAuthOptions
+>(async ({ org, prisma, role }, { searchParams }) => {
+    const params = await searchParams;
+    const page = pageSchema.safeParse(params.page).data ?? 1;
+    const search = z.string().optional().safeParse(params.search).data?.trim() ?? "";
+    const status = statusSchema.safeParse(params.status).data ?? "all";
+    const sortBy = sortBySchema.safeParse(params.sortBy).data ?? "name";
+    const sortOrder = sortOrderSchema.safeParse(params.sortOrder).data ?? "asc";
+    const skip = (page - 1) * DEFAULT_PAGE_SIZE;
+    const orderBy = sortBy === "indexedAt"
+        ? [{ indexedAt: sortOrder }, { id: "asc" as const }]
+        : [{ displayName: sortOrder }, { id: "asc" as const }];
+    const failedJobIds = status === "all"
+        ? []
+        : await getBullMQClient().getFailedJobIds(REPO_INDEX_QUEUE);
+    const where: Prisma.RepoWhereInput = {
+        orgId: org.id,
+        ...(search
+            ? {
+                  displayName: {
+                      contains: search,
+                      mode: "insensitive" as const,
+                  },
+              }
+            : {}),
+        ...(status === "all"
+            ? {}
+            : {
+                  latestIndexingJobId: { in: failedJobIds },
+                  indexedAt: status === "failed" ? null : { not: null },
+        }),
+    };
 
-    // Parse pagination parameters with defaults
-    const page = numberSchema.safeParse(params.page).data ?? 1;
-    const pageSize = numberSchema.safeParse(params.pageSize).data ?? DEFAULT_PAGE_SIZE;
-
-    // Parse filter parameters
-    const search = z.string().optional().safeParse(params.search).data ?? '';
-    const status = z.enum(['all', 'none', 'COMPLETED', 'IN_PROGRESS', 'PENDING', 'FAILED']).safeParse(params.status).data ?? 'all';
-    const sortBy = z.enum(['displayName', 'indexedAt']).safeParse(params.sortBy).data ?? undefined;
-    const sortOrder = z.enum(['asc', 'desc']).safeParse(params.sortOrder).data ?? 'asc';
-
-    // Calculate skip for pagination
-    const skip = (page - 1) * pageSize;
-
-    const _result = await getRepos({
-        skip,
-        take: pageSize,
-        search,
-        status,
-        sortBy,
-        sortOrder,
-    });
-    if (isServiceError(_result)) {
-        throw new ServiceErrorException(_result);
+    const [repos, totalCount] = await Promise.all([
+        prisma.repo.findMany({
+            where,
+            orderBy,
+            skip,
+            take: DEFAULT_PAGE_SIZE,
+            select: {
+                id: true,
+                name: true,
+                displayName: true,
+                indexedAt: true,
+                indexedCommitHash: true,
+                latestIndexingJobId: true,
+                imageUrl: true,
+                webUrl: true,
+                external_codeHostType: true,
+            },
+        }),
+        prisma.repo.count({
+            where,
+        }),
+    ]);
+    const latestJobIds = repos.flatMap((repo) =>
+        repo.latestIndexingJobId ? [repo.latestIndexingJobId] : [],
+    );
+    let latestJobs = new Map<string, WorkloadJob<"repo-index"> | null>();
+    try {
+        latestJobs = await getBullMQClient().getJobs(
+            REPO_INDEX_QUEUE,
+            latestJobIds,
+        );
+    } catch (error) {
+        console.error("Failed to load latest repository indexing jobs", error);
     }
 
-    const { repos, totalCount, stats } = _result;
-
     return (
-        <>
-            <div className="mb-6">
-                <h1 className="text-3xl font-semibold">Repositories</h1>
-                <p className="text-muted-foreground mt-2">View and manage your code repositories and their indexing status.</p>
+        <div className="flex flex-col h-full max-w-6xl mx-auto p-4">
+            <div className="flex flex-col gap-4 p-6 pb-2 flex-shrink-0">
+                <h3 className="text-lg font-medium">Repositories</h3>
             </div>
-            <ReposTable
-                data={repos.map((repo) => ({
-                    id: repo.id,
-                    name: repo.name,
-                    displayName: repo.displayName ?? repo.name,
-                    isArchived: repo.isArchived,
-                    isPublic: repo.isPublic,
-                    indexedAt: repo.indexedAt,
-                    createdAt: repo.createdAt,
-                    webUrl: repo.webUrl,
-                    imageUrl: repo.imageUrl,
-                    latestJobStatus: repo.latestIndexingJobStatus,
-                    isFirstTimeIndex: repo.indexedAt === null,
-                    codeHostType: repo.external_codeHostType,
-                    indexedCommitHash: repo.indexedCommitHash,
-                }))}
-                currentPage={page}
-                pageSize={pageSize}
-                totalCount={totalCount}
-                initialSearch={search}
-                initialStatus={status}
-                initialSortBy={sortBy}
-                initialSortOrder={sortOrder}
-                stats={stats}
-            />
-        </>
-    )
-}
-
-interface GetReposParams {
-    skip: number;
-    take: number;
-    search: string;
-    status: 'all' | 'none' | 'COMPLETED' | 'IN_PROGRESS' | 'PENDING' | 'FAILED';
-    sortBy?: 'displayName' | 'indexedAt';
-    sortOrder: 'asc' | 'desc';
-}
-
-const getRepos = async ({ skip, take, search, status, sortBy, sortOrder }: GetReposParams) => sew(() =>
-    withOptionalAuth(async ({ prisma }) => {
-        const whereClause: Prisma.RepoWhereInput = {
-            ...(search ? {
-                displayName: { contains: search, mode: 'insensitive' },
-            } : {}),
-            latestIndexingJobStatus:
-                status === 'all' ? undefined :
-                status === 'none' ? null :
-                status
-        };
-
-        // Build orderBy clause based on sortBy and sortOrder
-        const orderByClause: Prisma.RepoOrderByWithRelationInput = {};
-
-        if (sortBy === 'displayName') {
-            orderByClause.displayName = sortOrder === 'asc' ? 'asc' : 'desc';
-        } else if (sortBy === 'indexedAt') {
-            orderByClause.indexedAt = sortOrder === 'asc' ? 'asc' : 'desc';
-        } else {
-            // Default to displayName asc
-            orderByClause.displayName = 'asc';
-        }
-
-        const repos = await prisma.repo.findMany({
-            skip,
-            take,
-            where: whereClause,
-            orderBy: orderByClause,
-        });
-
-        // Calculate total count using the filtered where clause
-        const totalCount = await prisma.repo.count({
-            where: whereClause
-        });
-
-        // Status stats
-        const [
-            numCompleted,
-            numFailed,
-            numPending,
-            numInProgress,
-            numNoJobs
-        ] = await Promise.all([
-            prisma.repo.count({
-                where: {
-                    ...whereClause,
-                    latestIndexingJobStatus: RepoIndexingJobStatus.COMPLETED,
-                }
-            }),
-            prisma.repo.count({
-                where: {
-                    ...whereClause,
-                    latestIndexingJobStatus: RepoIndexingJobStatus.FAILED,
-                }
-            }),
-            prisma.repo.count({
-                where: {
-                    ...whereClause,
-                    latestIndexingJobStatus: RepoIndexingJobStatus.PENDING,
-                }
-            }),
-            prisma.repo.count({
-                where: {
-                    ...whereClause,
-                    latestIndexingJobStatus: RepoIndexingJobStatus.IN_PROGRESS,
-                }
-            }),
-            prisma.repo.count({
-                where: {
-                    ...whereClause,
-                    latestIndexingJobStatus: null,
-                }
-            }),
-        ])
-
-        return {
-            repos,
-            totalCount,
-            stats: {
-                numCompleted,
-                numFailed,
-                numPending,
-                numInProgress,
-                numNoJobs,
-            }
-        };
-    }));
+            <div className="px-6">
+                <ReposTable
+                    data={repos.map((repo) => ({
+                        id: repo.id,
+                        name: repo.name,
+                        displayName: repo.displayName,
+                        indexedAt: repo.indexedAt,
+                        indexedCommitHash: repo.indexedCommitHash,
+                        latestJob: repo.latestIndexingJobId
+                            ? latestJobs.get(repo.latestIndexingJobId) ?? null
+                            : null,
+                        imageUrl: repo.imageUrl,
+                        webUrl: repo.webUrl,
+                        codeHostType: repo.external_codeHostType,
+                    }))}
+                    currentPage={page}
+                    pageSize={DEFAULT_PAGE_SIZE}
+                    totalCount={totalCount}
+                    canRetry={role === OrgRole.OWNER}
+                    sortBy={sortBy}
+                    sortOrder={sortOrder}
+                />
+            </div>
+        </div>
+    );
+}, { allowAnonymous: true });
